fix: resolve issues
