+++ apps/docs/pages/getting-started/dependencies.mdx
@@ -3,6 +3,7 @@
 | Dependency                     | Used for                                                             |
 | ------------------------------ | -------------------------------------------------------------------- |
 | @bitcoinerlab/secp256k1        | ECC library (used with bitcoinjs-lib to decode transactions)         |
+| @react-native-community/netinfo| Network Information                                                  |
 | @react-native-community/slider | Slider                                                               |
 | @shopify/react-native-skia     | Utxo Bubbles                                                         |
 | assert                         | Assert nodejs core module (required by varuint and/or bitcoinjs)     |
