+++ apps/mobile/app/(authenticated)/index.tsx
@@ -1,6 +1,5 @@
-import NetInfo from '@react-native-community/netinfo'
 import { Stack, useRouter } from 'expo-router'
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import { useState } from 'react'
 import { ScrollView, View } from 'react-native'
 import { useShallow } from 'zustand/react/shallow'
 
@@ -13,13 +12,13 @@ import SSAccountCard from '@/components/SSAccountCard'
 import SSButton from '@/components/SSButton'
 import SSSeparator from '@/components/SSSeparator'
 import SSText from '@/components/SSText'
+import useVerifyConnection from '@/hooks/useVerifyConnection'
 import SSHStack from '@/layouts/SSHStack'
 import SSMainLayout from '@/layouts/SSMainLayout'
 import SSVStack from '@/layouts/SSVStack'
 import { t } from '@/locales'
 import { useAccountBuilderStore } from '@/store/accountBuilder'
 import { useAccountsStore } from '@/store/accounts'
-import { useBlockchainStore } from '@/store/blockchain'
 import { Colors } from '@/styles'
 import { sampleSignetWalletSeed } from '@/utils/samples'
 
@@ -45,10 +44,6 @@ export default function AccountList() {
       ])
     )
 
-  const [network, url, timeout] = useBlockchainStore(
-    useShallow((state) => [state.network, state.url, state.timeout * 1000])
-  )
-
   const [loadingWallet, setLoadingWallet] = useState(false)
 
   async function loadSampleSignetWallet() {
@@ -70,98 +65,8 @@ export default function AccountList() {
     }
   }
 
-  const isConnectionAvailable = useRef<boolean | null>(false)
-  const [connectionState, setConnectionState] = useState<boolean>(false)
-  const connectionString = useMemo(() => {
-    return network + ' - ' + url
-  }, [network, url])
-
-  const verifyUrl = useMemo(() => {
-    const urlObj = new URL(url)
-    if (urlObj.protocol === 'ssl:') {
-      const modifiedUrl = new URL(url.replace('ssl://', 'https://'))
-      modifiedUrl.port = ''
-      modifiedUrl.pathname = '/signet/api/v1/difficulty-adjustment'
-      return modifiedUrl.toString()
-    } else {
-      urlObj.pathname = '/api/v1/difficulty-adjustment'
-    }
-    return urlObj.toString()
-  }, [url])
-
-  const isPrivateConnection = useMemo(() => {
-    if (
-      url === 'ssl://mempool.space:60602' ||
-      url === 'https://mutinynet.com/api'
-    ) {
-      return false
-    }
-    return true
-  }, [url])
-
-  const verifyConnection = useCallback(async () => {
-    if (!isConnectionAvailable.current) {
-      setConnectionState(false)
-      return
-    }
-    try {
-      const timeoutPromise = new Promise((_, reject) =>
-        setTimeout(() => reject(new Error('timeout')), timeout)
-      )
-      const fetchPromise = fetch(verifyUrl)
-      const response = (await Promise.race([
-        timeoutPromise,
-        fetchPromise
-      ])) as Response
-      if (!isConnectionAvailable.current) {
-        setConnectionState(false)
-        return
-      }
-      if (response.ok) {
-        setConnectionState(true)
-      } else {
-        setConnectionState(false)
-      }
-    } catch (_) {
-      setConnectionState(false)
-    }
-  }, [timeout, verifyUrl])
-
-  const checkConnection = useCallback(async () => {
-    const state = await NetInfo.fetch()
-    isConnectionAvailable.current = state.isConnected
-  }, [])
-
-  useEffect(() => {
-    ;(async () => {
-      await checkConnection()
-      verifyConnection()
-    })()
-    const timerId = setInterval(() => {
-      verifyConnection()
-    }, 60000)
-
-    const unsubscribe = NetInfo.addEventListener((state) => {
-      if (
-        isConnectionAvailable.current !== state.isConnected &&
-        state.isConnected !== null
-      ) {
-        isConnectionAvailable.current = state.isConnected
-        if (state.isConnected) {
-          setTimeout(verifyConnection, 5000)
-        } else {
-          verifyConnection()
-        }
-      } else {
-        isConnectionAvailable.current = state.isConnected
-      }
-    })
-
-    return () => {
-      unsubscribe()
-      clearInterval(timerId)
-    }
-  }, [checkConnection, verifyConnection])
+  const [connectionState, connectionString, isPrivateConnection] =
+    useVerifyConnection()
 
   return (
     <>
