+++ apps/mobile/styles/colors.ts
@@ -17,9 +17,9 @@ export const gray = {
   700: '#363636',
   800: '#2D2D2D',
   850: '#242424',
-  870: '#212121',
+  875: '#212121',
   900: '#1A1A1A',
-  910: '#1B1B1B',
+  925: '#1B1B1B',
   950: '#131313'
 }
